refactor(scheduler): apply review feedback on group overrides
Rename the table and model to group_overrides so it can hold other
per-group attributes later, rename set to upsert, and require
max_concurrency > 0 since 0 means unlimited in the dequeue gate.
