+++ packages/scheduler/lib/models/tasks.integration.test.ts
@@ -5,7 +5,7 @@ import { nanoid } from '@nangohq/utils';
 import { getTestDbClient } from '../db/helpers.test.js';
 import { isDuplicateTaskNameError } from '../errors.js';
 import { taskStates } from '../types.js';
-import * as concurrencyOverrides from './concurrencyOverrides.js';
+import * as groupOverrides from './groupOverrides.js';
 import * as tasks from './tasks.js';
 
 import type { Task, TaskState } from '../types.js';
@@ -383,7 +383,7 @@ describe('Task', () => {
         });
         it('should reflect a concurrency override stamped at create time', async () => {
             // the override is set before the tasks are created, so it is stamped onto them
-            (await concurrencyOverrides.set(db, { groupKey: 'sync:environment:1', maxConcurrency: 2 })).unwrap();
+            (await groupOverrides.upsert(db, { groupKey: 'sync:environment:1', maxConcurrency: 2 })).unwrap();
             for (let i = 0; i < 3; i++) {
                 await createTask(db, { groupKey: 'sync:environment:1', groupMaxConcurrency: 5 });
             }
@@ -393,14 +393,14 @@ describe('Task', () => {
     describe('concurrency overrides', () => {
         it('stamps the override onto the task at create time', async () => {
             const groupKey = nanoid();
-            (await concurrencyOverrides.set(db, { groupKey, maxConcurrency: 2 })).unwrap();
+            (await groupOverrides.upsert(db, { groupKey, maxConcurrency: 2 })).unwrap();
 
             const task = await createTask(db, { groupKey, groupMaxConcurrency: 5 });
             expect(task.groupMaxConcurrency).toBe(2);
         });
         it('caps dequeue by an override set before the tasks are created', async () => {
             const groupKey = nanoid();
-            (await concurrencyOverrides.set(db, { groupKey, maxConcurrency: 2 })).unwrap();
+            (await groupOverrides.upsert(db, { groupKey, maxConcurrency: 2 })).unwrap();
             for (let i = 0; i < 5; i++) {
                 await createTask(db, { groupKey, groupMaxConcurrency: 5 });
             }
@@ -410,7 +410,7 @@ describe('Task', () => {
         });
         it('can raise concurrency above the default', async () => {
             const groupKey = nanoid();
-            (await concurrencyOverrides.set(db, { groupKey, maxConcurrency: 3 })).unwrap();
+            (await groupOverrides.upsert(db, { groupKey, maxConcurrency: 3 })).unwrap();
             for (let i = 0; i < 3; i++) {
                 await createTask(db, { groupKey, groupMaxConcurrency: 1 });
             }
@@ -424,17 +424,17 @@ describe('Task', () => {
             expect(before.groupMaxConcurrency).toBe(5);
 
             // setting the override only stamps tasks created afterwards
-            (await concurrencyOverrides.set(db, { groupKey, maxConcurrency: 2 })).unwrap();
+            (await groupOverrides.upsert(db, { groupKey, maxConcurrency: 2 })).unwrap();
             const after = await createTask(db, { groupKey, groupMaxConcurrency: 5 });
             expect(after.groupMaxConcurrency).toBe(2);
         });
         it('reverts to the default once the override is removed', async () => {
             const groupKey = nanoid();
-            (await concurrencyOverrides.set(db, { groupKey, maxConcurrency: 1 })).unwrap();
+            (await groupOverrides.upsert(db, { groupKey, maxConcurrency: 1 })).unwrap();
             const capped = await createTask(db, { groupKey, groupMaxConcurrency: 5 });
             expect(capped.groupMaxConcurrency).toBe(1);
 
-            (await concurrencyOverrides.remove(db, groupKey)).unwrap();
+            (await groupOverrides.remove(db, groupKey)).unwrap();
             const uncapped = await createTask(db, { groupKey, groupMaxConcurrency: 5 });
             expect(uncapped.groupMaxConcurrency).toBe(5);
         });
